We add and remove few of our custom hooks dynamically when user change the Payment Method on checkout page.
Before add our custom buttons in Order Edit page we try to check is this
order made via SafeCharge Paygate. If the order does not have our
specific meta fields we accept it is not made via SC and do not show SC
buttons.

In this version will be added new functionality - complete Refund,
when user change the status of the order to Refunded. In this case
the WC create complete refund.